Add files via upload
The "Material" folder contains a collection of UI screenshots from various versions of Microsoft Office applications, including Excel, Word, and PowerPoint from the years 2007, 2016, 2019, and 2021. These images are intended for use in generating and evaluating saliency heat maps. The purpose of this analysis is to assess the visual focus and user experience across different iterations of the MS Office interface.